move BIP32 example to own file
